Add Conditions
The Condition class has been added. It serves as a base class for
conditions that can be checked on SulfideElements. E.g. the
ExistCondition checks if a SulfideElement exists in the page.

SulfideElement has been moved to a proper class.
SulfideElement has a new method should that can be used to check if a
condition for the element is met within a given time. Method shouldExist
is now a shortcut for should(exist()).

SulfideElement class definition is passed as argument to Selectors and
Conditions.