Add SulfideElementCollection class and LengthCondition
The SulfideElementCollection represents a list of elements specified by
a selector (not an xpath!). The factory method $$ has been added to
Sulfide and the global namespace.

The LengthCondition can be used to test if an element list represented by
a SulfideElementCondition consists of a certain number of elements.

Example: $$('div').shouldHave(length(10))