Add support for NWJS apps
Adding an option to specify running an NWJS app.
Overrides the Target.page() function of puppeteer when running an NWJS
app to be able to get the page handle from a target.
'targetchanged' listener is added to $.open method when runing an NWJS
app to get the page handle.
Updated .eslintrc to support spread operator and dangling underscores.

TODO: sulfide should be able to handle multiple pages.